+++ client/src/components/Stores/StoreOptions.vue
@@ -5,20 +5,28 @@
   >
     <StoreInformationModal
       :is-jumper="isJumper"
+      :is-manager="isManager"
       :store-id="storeId"
       :may-edit-store="mayEditStore"
       :is-coordinator="isCoordinator"
       :is-verified="isVerified"
     />
     <button
-      v-if="teamConversationId != null && isUserInStore"
+      v-if="coordinatorConversationId && isManager"
+      type="button"
+      class="list-group-item list-group-item-action"
+      @click="openChat(coordinatorConversationId)"
+      v-text="$i18n('store.chat.coordinators')"
+    />
+    <button
+      v-if="teamConversationId && isUserInStore"
       type="button"
       class="list-group-item list-group-item-action"
       @click="openChat(teamConversationId)"
       v-text="$i18n('store.chat.team')"
     />
     <button
-      v-if="jumperConversationId != null && isUserInStore || isJumper"
+      v-if="jumperConversationId && isUserInStore || isJumper"
       type="button"
       class="list-group-item list-group-item-action"
       @click="openChat(jumperConversationId)"
@@ -58,26 +66,13 @@ export default {
     storeName: { type: String, required: true },
     fsId: { type: Number, required: true },
     mayLeaveStoreTeam: { type: Boolean, default: false },
-    teamConversationId: {
-      type: Number,
-      default: null,
-    },
-    jumperConversationId: {
-      type: Number,
-      default: null,
-    },
-    mayEditStore: {
-      type: Boolean,
-      default: null,
-    },
-    isCoordinator: {
-      type: Boolean,
-      default: null,
-    },
-    storeId: {
-      type: Number,
-      default: null,
-    },
+    coordinatorConversationId: { type: Number, default: null },
+    teamConversationId: { type: Number, default: null },
+    jumperConversationId: { type: Number, default: null },
+    mayEditStore: { type: Boolean, default: null },
+    isManager: { type: Boolean, default: null },
+    isCoordinator: { type: Boolean, default: null },
+    storeId: { type: Number, default: null },
     isUserInStore: { type: Boolean, default: false },
     isJumper: { type: Boolean, default: false },
     mayDoPickup: { type: Boolean, default: false },