+++ tests/unit/StoreGatewayTest.php
@@ -10,6 +10,7 @@ use Foodsharing\Modules\Core\DBConstants\Unit\UnitType;
 use Foodsharing\Modules\Region\DTO\MinimalRegionIdentifier;
 use Foodsharing\Modules\Store\DTO\Store;
 use Foodsharing\Modules\Store\StoreGateway;
+use Foodsharing\Modules\Store\StoreTransactions;
 use Foodsharing\Modules\Store\TeamStatus;
 
 class StoreGatewayTest extends Unit
@@ -358,16 +359,23 @@ class StoreGatewayTest extends Unit
         $this->tester->seeInDatabase('fs_betrieb', ['bezirk_id' => $newRegion['id'], 'id' => $this->store['id']]);
     }
 
+    public function testGetNoCoordinatorConversation(): void
+    {
+        $conversationId = $this->gateway->getStoreConversation($this->store['id'], StoreTransactions::CONVERSATION_TYPE_COORDINATOR);
+
+        $this->tester->assertEquals(0, $conversationId);
+    }
+
     public function testGetNoTeamConversation(): void
     {
-        $conversationId = $this->gateway->getBetriebConversation($this->store['id']);
+        $conversationId = $this->gateway->getStoreConversation($this->store['id'], StoreTransactions::CONVERSATION_TYPE_TEAM);
 
         $this->tester->assertEquals(0, $conversationId);
     }
 
     public function testGetNoSpringerConversation(): void
     {
-        $conversationId = $this->gateway->getBetriebConversation($this->store['id'], true);
+        $conversationId = $this->gateway->getStoreConversation($this->store['id'], StoreTransactions::CONVERSATION_TYPE_SPRINGER);
 
         $this->tester->assertEquals(0, $conversationId);
     }